restrict usage of personality to Linux
Other OSes, e.g. macOS, are missing this functionality. For now we just live
with ASLR being left enabled in such environments. However, there may be a way
to pass the undocumented _POSIX_SPAWN_DISABLE_ASLR through to
g_spawn_async_with_pipes to achieve this in future. This is how GDB on macOS
disables ASLR.